Update to v5.8.3 and use new install format
Ghost has significantly changed their file structure, which means
the scripts had to be changed around quite a bit as well. We are
also using the readymade `yarn setup` tool to set up the Ghost
installation instead of manually running each command. Significant
changes are:

  - Config and content now lives under ./ghost/core instead of
    being directly in the project root. While updating, for
    backward compatibility, we're checking for the existence
    of the old folder as well
  - The admin interface is bundled along with the main Ghost
    package, so there is no need to download it separately
    (Casper is still being downloaded separately though)
  - We are using `yarn start` instead of knex, grunt, etc.
    (`yarn start` automatically calls those commands for us
    behind the scenes)
  - Similarly, we are now running `yarn start` to execute the
    Ghost process in the systemd config, instead of manually
    running `index.js` as we were doing earlier.

Fresh installs work well on this commit, though upgrades are yet
to be tested!