Increase yarn network timeout to 10 minutes
We often run into the error, `An unexpected error occurred:
"https://registry.yarnpkg.com/date-fns/-/date-fns-2.29.1.tgz:
ESOCKETTIMEDOUT".` This seems to be a common problem with the
date-fns package, as well as other large packages installed with
older versions of Yarn. See for example the following GitHub
issues:

https://github.com/date-fns/date-fns/issues/1004
https://github.com/yarnpkg/yarn/issues/8242

The problem seems to be related to slow hard-disk access on the
client server/VPS - so it's actually the storage connection and
not the Internet connection, that's timing out.

A common workaround for this is to increase the network timeout
for Yarn long enough for the build will complete, which is what
we're doing in this case since it's likely to affect a lot of the
slower setups (including, as it seems, the YunoHost CI!). For
everyone else, the worst case is that they'll have to wait 10
minutes before they realise there's actually something wrong with
the network.